python/requirements.txt
bitcoin==1.1.42
ethereum==2.3.1
pysha3==1.0.2
tornado==5.0.2
web3==4.2.0
eth-rlp==0.1.2
rlp==0.6.0
eth-account==0.2.2
